+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Add `UpdatesController#webhook_request` that is set to `ActionDispatch::Request`
+  in webhook mode.
 - Update to Bot API 6.8
 - Added `chat_join_request` to PAYLOAD_TYPES